Updated License
GPL restricts derivative or redistributed code to be used under the same licence, which is a restriction that I don't care for. Reading up, it seems that the MIT License is more appropriate.

Do whatever the heck you want with this, just be sure to credit me where ever appropriate.